+++ src/hooks/api/user/useLoginQuery.ts
@@ -1,6 +1,30 @@
+import { UserType } from "@/types/global";
 import { instance } from "@/utils/instance";
 import { useMutation } from "@tanstack/react-query";
-import { LoginRequest, LoginResponse } from "@/types/user";
+
+export interface LoginRequest {
+  email: string;
+  password: string;
+}
+
+export interface LoginResponse {
+  item: {
+    token: string;
+    user: {
+      item: {
+        id: string;
+        email: string;
+        type: UserType;
+        name: string;
+        phone: string;
+        address: string;
+        bio: string;
+      };
+      href: string;
+    };
+  };
+  links: [];
+}
 
 const postLogin = async ({ email, password }: LoginRequest): Promise<LoginResponse> => {
   const response = await instance.post("/token", { email, password });
