+++ src/hooks/api/user/usePutMyInfoQuery.ts
@@ -1,9 +1,18 @@
-import { PutMyInfoRequest } from "@/types/user";
+import { SeoulAddress } from "@/types/global";
 import { instance } from "@/utils/instance";
 import { useMutation } from "@tanstack/react-query";
-import { User } from "@/types/user";
+import { GetMyInfoResponse } from "./useGetMyInfoQuery";
 
-const putMyInfo = async ({ userId, data }: { userId: string; data: PutMyInfoRequest }): Promise<User> => {
+export interface PutMyInfoRequest {
+  name: string;
+  phone: string;
+  address: SeoulAddress;
+  bio: string;
+}
+
+export type PutMyInfoResponse = GetMyInfoResponse;
+
+const putMyInfo = async ({ userId, data }: { userId: string; data: PutMyInfoRequest }): Promise<PutMyInfoResponse> => {
   const response = await instance.put(`/users/${userId}`, data);
   return response.data;
 };
